Use random IP addresses as default Diag terminal
As pointed in [1], SAP systems that doesn't implement SAP Note 1497445 [2] uses the terminal name as provided by the client to identify the source of the connection, or fallbacks to using the IP address if the terminal name can't be resolved. Using a random IP address makes unreliable the source of the conection in the Security Audit Log (SAL).

[1] https://www.troopers.de/troopers14-presentations/index.html#SAPAntiForensics
[2] http://service.sap.com/sap/support/notes/1497445