Added SAPRoutedStreamSocket
Added a Stream Socket implementation that wraps the NIStreamSocket and allows request a connection through a SAP Router to a given target. After the route request is performed, the Stream Socket is used as a regular one with the target host/port. This new SAPRoutedStreamSocket is used in the router_scanner example script.