Adding new script to fingerprint SAP Router
This is an experiment I've been working on some time ago. When SAP Router is not configure to hide error details, it throws away lots of interesting information we can use to fingerprint different versions. This is an experiment no having a fingerprint tool and an initial database filled up with 8 different versions.